requirements.txt
absl-py==1.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
cycler==0.11.0
demjson==2.2.4
flatbuffers==1.12
flwr==0.19.0
fonttools==4.33.3
gast==0.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h11==0.13.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
iterators==0.0.2
jtutils==0.0.8
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
leven==1.0.4
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
nose==1.3.7
numpy==1.20.3
oauthlib==3.2.0
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.9
python-csv==0.0.13
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
xlrd==2.0.1
xmltodict==0.13.0
zipp==3.8.0
